refactor(compiler): change compiler exported symbols
